Update NBA_Stat_Tweeter_Bot.py
Accesses player stats data stored in AWS RDS via mySQL connection. Then scans Twitter mention tweets for names of current NBA players, replying to those tweets with the respective values for the number of minutes played this season for the players that were identified. 
This code has been uploaded on a crontab on an AWS EC2 instance, being executed every 10 minutes. 

In the future, as I'm able to grow what types of statistics are stored in the database, I may increase the scope of what type of information the reply tweet contains.